wayfarer/anne x/Nomad/requirements.txt
fastapi
uvicorn[standard]
pydantic
PyYAML
mavsdk
aiofiles
python-multipart
pymavlink
pytest
pyserial
paho-mqtt